refactor: Improve UserProfile and UserProfileDetail components
This commit refactors the UserProfile and UserProfileDetail components to improve their functionality and user experience. It adds a console.log statement in the handleSelectedSettingOption function in UserProfile.vue and removes the import of SettingProfileOptions from UserProfile.enum in UserProfileDetail.vue. It also updates the watch function in UserProfileDetail.vue to check for the id property of props.settingOption instead of the name property. These changes enhance the UserProfile and UserProfileDetail components.