refactor: Improve status display in AssignmentDetails and Dashboard
Update the VaBadge component in AssignmentDetails.vue and Dashboard.vue to improve the display of assignment status. This change ensures that the status is shown more prominently and with appropriate colors. Additionally, the getMarkAssignmentStatusText and getMarkAssignmentStatusColor functions have been added to provide localized status text and color based on the assignment status.

See merge request guardian-exam-online/geo-vue-managerment-fe!168